requirement.txt
torch
Pillow
opencv-python
matplotlib
tqdm
torchvision
segmentation-models-pytorch
PyQt6